dotnet-sdk: Revert to LLVM 15 for now
With LLVM 20 out and most projects already migrated to it, keeping an
old LLVM around is costly. LLVM 15 is still needed for PostgreSQL so
it’s here to stay and is a safer bet if you want to limit your used disk
space by having as few installed slots as possible.

Signed-off-by: Morgane Glidic <sardemff7+git@sardemff7.net>